test: move duplicate test checks into common config class
not need to have different checks when we're using the ProcessStore
interface (which is already abstracting the difference between OS and
ES)
