+++ tasklist/qa/integration-tests/src/test/java/io/camunda/tasklist/util/TestChecks.java
@@ -8,17 +8,74 @@
 package io.camunda.tasklist.util;
 
 import static io.camunda.tasklist.util.TestCheck.FORM_EXISTS_CHECK;
+import static io.camunda.tasklist.util.TestCheck.PROCESS_IS_DELETED_CHECK;
+import static io.camunda.tasklist.util.TestCheck.PROCESS_IS_DEPLOYED_CHECK;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.camunda.tasklist.exceptions.NotFoundException;
+import io.camunda.tasklist.exceptions.TasklistRuntimeException;
 import io.camunda.tasklist.store.FormStore;
+import io.camunda.tasklist.store.ProcessStore;
+import io.camunda.webapps.schema.entities.ProcessEntity;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 @Configuration
 public class TestChecks {
+
+  @Autowired private ProcessStore processStore;
+
   @Autowired private FormStore formStore;
+
+  /** Checks whether the process of given args[0] processId (Long) is deployed. */
+  @Bean(name = PROCESS_IS_DEPLOYED_CHECK)
+  public TestCheck getProcessIsDeployedCheck() {
+    return new TestCheck() {
+      @Override
+      public String getName() {
+        return PROCESS_IS_DEPLOYED_CHECK;
+      }
+
+      @Override
+      public boolean test(final Object[] objects) {
+        assertThat(objects).hasSize(1);
+        assertThat(objects[0]).isInstanceOf(String.class);
+        final String processId = (String) objects[0];
+        try {
+          final ProcessEntity process = processStore.getProcess(processId);
+          return process != null;
+        } catch (final TasklistRuntimeException ex) {
+          return false;
+        }
+      }
+    };
+  }
+
+  @Bean(name = PROCESS_IS_DELETED_CHECK)
+  public TestCheck getProcessIsDeletedCheck() {
+    return new TestCheck() {
+      @Override
+      public String getName() {
+        return PROCESS_IS_DELETED_CHECK;
+      }
+
+      @Override
+      public boolean test(final Object[] objects) {
+        assertThat(objects).hasSize(1);
+        assertThat(objects[0]).isInstanceOf(String.class);
+        final String processId = (String) objects[0];
+        try {
+          processStore.getProcess(processId);
+          return false;
+        } catch (final NotFoundException nfe) {
+          return true;
+        } catch (final TasklistRuntimeException ex) {
+          return false;
+        }
+      }
+    };
+  }
 
   @Bean(name = FORM_EXISTS_CHECK)
   public TestCheck getFormExistsCheck() {
